+++ src/Semver/Properties/AssemblyInfo.cs
