+++ test/Semver.Test/Properties/AssemblyInfo.cs
